Fix variable type

+++ src/chat/encryption/lime-v2.cpp
@@ -351,7 +351,7 @@ bool LimeV2::encryptionEnabledForFileTransfer (const shared_ptr<AbstractChatRoom
 }
 
 void LimeV2::generateFileTransferKey (const shared_ptr<AbstractChatRoom> &chatRoom, const shared_ptr<ChatMessage> &message) {
-	int FILE_TRANSFER_KEY_SIZE = 32;
+	size_t FILE_TRANSFER_KEY_SIZE = 32;
 	char keyBuffer [FILE_TRANSFER_KEY_SIZE];// temporary storage of generated key: 192 bits of key + 64 bits of initial vector
 	// generate a random 192 bits key + 64 bits of initial vector and store it into the file_transfer_information->key field of the msg
     sal_get_random_bytes((unsigned char *)keyBuffer, FILE_TRANSFER_KEY_SIZE);
